Dynatrace/requirements.txt
fastapi
requests
pydantic
python-dotenv
jinja2
